% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_input}
\alias{.check_input}
\title{Check input to functions}
\usage{
.check_input(x)
}
\arguments{
\item{x}{TODO A Seurat or SingleCellExperiment object}
}
\value{
Invisible NULL if no error is triggered
}
\description{
Internal function
}
