% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_assay_specific_features.R
\name{find_assay_specific_features}
\alias{find_assay_specific_features}
\title{Title}
\usage{
find_assay_specific_features(
  seurat_obj,
  assay_var = "assay",
  n_features = 5000
)
}
\arguments{
\item{n_features}{}
}
\value{

}
\description{
Title
}
