% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{integrate_assays}
\alias{integrate_assays}
\title{Title}
\usage{
integrate_assays(
  seurat_obj,
  assay_specific = TRUE,
  assay_var = "assay",
  shared_hvg,
  n_dim = 30
)
}
\arguments{
\item{n_dim}{}
}
\value{

}
\description{
Title
}
