% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays_atac.R
\name{integrate_assays_atac}
\alias{integrate_assays_atac}
\title{Title}
\usage{
integrate_assays_atac(
  seurat_obj,
  n_dim = 40,
  assay_use,
  group_by_vars,
  reduction
)
}
\arguments{
\item{reduction}{}
}
\value{

}
\description{
Title
}
