% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLOcatoR-pkg.R
\docType{package}
\name{SLOcatoR-pkg}
\alias{SLOcatoR-pkg}
\title{SLOcatoR}
\description{
\code{SLOcatoR} is a Bioconductor package that provides TODO TODO
}
\details{
TODO TODO something more

Can contain quite a lot of things in here, remove once done
}
