% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_training_and_test_sets.R
\name{split_training_and_test_sets}
\alias{split_training_and_test_sets}
\title{Title}
\usage{
split_training_and_test_sets(
  seurat_obj,
  split_var,
  referece_label,
  query_label,
  reduction = "harmony",
  n_dims = 30
)
}
\arguments{
\item{n_dims}{}
}
\value{

}
\description{
Title
}
