% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_umap_coords.R
\name{transfer_umap_coords}
\alias{transfer_umap_coords}
\title{Title}
\usage{
transfer_umap_coords(
  seurat_obj,
  training_set,
  test_set,
  umap1_var,
  umap2_var,
  k
)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Title
}
