% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_assay_specific_features.R
\name{find_assay_specific_features}
\alias{find_assay_specific_features}
\title{Find Assay-specific Highly Variable Features}
\usage{
find_assay_specific_features(
  seurat_obj,
  assay_var = "assay",
  n_features = 5000
)
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{assay_var}{a character string specifying which variable in
seurat_obj@meta.data contains the assay for each cell}

\item{n_features}{number of highly variable features to find per assay}
}
\value{
A character vector with all the highly variable features shared
across assays
}
\description{
scRNA-seq and snRNA-seq integrate very poorly (Mereu & Lafzi, 2020). This is
caused by the fact that both techniques sample transcripts from different
cellular compartments. Similarly, we have obtained a poor mixability after
integrating CITE-seq and scRNA-seq. To overcome it, we propose performing
feature extraction for each assay independently, and then taking the
intersection. In this setting, assay-dependent variable features are discarded,
and batch effects are largely mitigated.
}
\examples{
## TODO TODO

}
