% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_k.R
\name{find_optimal_k}
\alias{find_optimal_k}
\title{Title}
\usage{
find_optimal_k(
  seurat_obj,
  training_set,
  response_var,
  ks = c(2, 4, 6, 8, 16, 32, 64, 128, 256),
  return_plot = TRUE
)
}
\arguments{
\item{return_plot}{}
}
\value{

}
\description{
Title
}
