% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_umap_coords.R
\name{transfer_umap_coords}
\alias{transfer_umap_coords}
\title{Transfer UMAP coordinates from training set to test set using KNN regression}
\usage{
transfer_umap_coords(
  seurat_obj,
  training_set,
  test_set,
  umap1_var,
  umap2_var,
  k
)
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{training_set}{a matrix of cells x features. Usually corresponds to the
"training_set" element in the list obtained with split_training_and_test_sets}

\item{test_set}{a matrix of cells x features. Usually corresponds to the
"test_set" element in the list obtained with split_training_and_test_sets}

\item{umap1_var}{a character string specifying which variable in
seurat_obj@meta.data contains the UMAP1 coordinates}

\item{umap2_var}{a character string specifying which variable in
seurat_obj@meta.data contains the UMAP2 coordinates}

\item{k}{numeric specifying the size of the neighborhood}
}
\value{
A data.frame that contains the predicted UMAP1 and UMAP2 coords for
each cell.
}
\description{
Transfer UMAP coordinates from training set to test set using KNN regression
}
\examples{

## TODO TODO

}
